Moved fonts to a Fonts folder and updated scss files
I removed the Roboto font entirely, deciding to use a Google Font import
because the Roboto import I was trying to do wasn't working.  I then
moved all the font files to the src/fonts directory.

The Gulpfile will need to be updated for this change to handle the
importing of the font files.